+++ CHANGELOG.md
@@ -6,9 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [0.3.5]
 
+### Added
+
+- Added a new `VibThermoSchema` that merges the `VibSchema` and `ThermoSchema` as a convenience
+
 ### Changed
 
-- Changed default `SCRATCH_DIR` from `Path.cwd() / ".scratch"` to `Path("~/.scratch")
+- The output of recies no longer automatically removes blank lists/dicts/sets to maintain a consistent schema.
+- The tblite `freq_job` and lj `freq_job` now use `VibThermoSchema` as the output
+- Changed default `SCRATCH_DIR` from `Path.cwd() / ".scratch"` to `Path("~/.scratch")`
 
 ## [0.3.4]